Added homescreen and extracted to game
